+++ plugins/kernel_v2/config/ssh.rb
@@ -12,6 +12,7 @@ module VagrantPlugins
       attr_accessor :shell
       attr_accessor :proxy_command
       attr_accessor :pty
+      attr_accessor :sudo_command
 
       attr_reader :default
 
@@ -25,6 +26,7 @@ module VagrantPlugins
         @proxy_command = UNSET_VALUE
         @pty           = UNSET_VALUE
         @shell         = UNSET_VALUE
+        @sudo_command  = UNSET_VALUE
 
         @default    = SSHConnectConfig.new
       end
@@ -46,6 +48,10 @@ module VagrantPlugins
         @proxy_command = nil if @proxy_command == UNSET_VALUE
         @pty        = false if @pty == UNSET_VALUE
         @shell      = "bash -l" if @shell == UNSET_VALUE
+
+        if @sudo_command == UNSET_VALUE
+          @sudo_command = "sudo -E -H %c"
+        end
 
         @default.username = "vagrant" if @default.username == UNSET_VALUE
         @default.port     = @guest_port if @default.port == UNSET_VALUE